// Front-end GIROHOSTING/Giro/src/app/models/userFullData.model.ts
//Model de totes les dades de l'usuari
export class UserFullData
{
    id: any;
    rowid: any;
    nom: any;
    name_alias: any;
    email: any;
    password: any;
    phone: any;
    address: any;
    town: any;
    fk_pays: any;
    fk_departement: any;
    zip: any;
    siren: any;
    url: any;
}